feat: update auth flow
