+++ src/hooks.server.ts
@@ -68,10 +68,10 @@ const authGuard: Handle = async ({ event, resolve }) => {
   event.locals.user = user
 
   if (!event.locals.session && event.url.pathname.startsWith('/app')) {
-    redirect(303, '/auth')
+    redirect(303, '/auth/login')
   }
 
-  if (event.locals.session && event.url.pathname === '/auth') {
+  if (event.locals.session && event.url.pathname.includes('/auth')) {
     redirect(303, '/app')
   }
