--- src/routes/auth/+page.svelte
@@ -1,12 +1,0 @@
-<form method="POST" action="?/login" class="flex flex-col gap-2">
-  <label for="email">Email</label>
-  <input name="email" type="email" required class="border border-border rounded-md p-2"/>
-  <label for="firstName">First Name</label>
-  <input name="firstName" type="text" required class="border border-border rounded-md p-2"/>
-  <label for="lastName">Last Name</label>
-  <input name="lastName" type="text" required class="border border-border rounded-md p-2"/>
-  <label for="password">Password</label>
-  <input name="password" type="password" required class="border border-border rounded-md p-2"/>
-  <button type="submit">Login</button>
-  <button formaction="?/signup">Sign up</button>
-</form>
