+++ src/routes/auth/login/+page.server.ts
@@ -1,21 +1,18 @@
-// import { redirect } from '@sveltejs/kit';
-// import { ClientResponseError } from 'pocketbase';
-// import type { Actions } from './$types';
+import { redirect } from '@sveltejs/kit';
+import type { Actions } from './$types';
 
-// export const actions: Actions = {
-// 	default: async ({ request, locals }) => {
-//     const body = Object.fromEntries(await request.formData());
+export const actions: Actions = {
+    default: async ({ request, locals: { supabase } }) => {
+        const formData = await request.formData()
+        const email = formData.get('email') as string
+        const password = formData.get('password') as string
 
-//     try {
-//       await locals.pb.collection('users').authWithPassword(body.email, body.password);
-//       return redirect(303, '/app');
-//     } catch (error) {
-//       if (error instanceof ClientResponseError) {
-//         return {error: error.response.message};
-//       } else {
-//         console.error(error);
-//         return redirect(303, '/auth/error');
-//       }
-//     }
-// 	}
-// };
+        const { error } = await supabase.auth.signInWithPassword({ email, password })
+        if (error) {
+            console.error(error)
+            redirect(303, '/auth/error')
+        } else {
+            redirect(303, '/app')
+        }
+    }
+};
