+++ src/routes/auth/signup/+page.server.ts
@@ -1,18 +1,22 @@
-// import { redirect } from '@sveltejs/kit';
+import { redirect } from '@sveltejs/kit';
 
-// import type { Actions } from './$types';
+import type { Actions } from './$types';
 
-// export const actions: Actions = {
-// 	default: async ({ request, locals }) => {
-//     const body = Object.fromEntries(await request.formData());
+export const actions: Actions = {
+    default: async ({ request, locals: { supabase } }) => {
+        const formData = await request.formData()
+        const firstName = formData.get('firstName') as string
+        const lastName = formData.get('lastName') as string
+        const email = formData.get('email') as string
+        const password = formData.get('password') as string
 
-//     try {
-//       await locals.pb.collection('users').create(body);
-//       await locals.pb.collection('users').authWithPassword(body.email, body.password);
-//     } catch (error) {
-//       console.error(error);
-//       return redirect(303, '/auth/error');
-//     }
-//     return redirect(303, '/app');
-// 	}
-// };
+        const { error } = await supabase.auth.signUp({ email, password, options: { data: { firstName, lastName } } })
+
+        if (error) {
+            console.error(error)
+            redirect(303, '/auth/error')
+        }
+
+        redirect(303, '/app')
+    }
+};
